+++ hydromt/data_adapter/geodataframe.py
@@ -14,6 +14,7 @@ from pystac import Item as StacItem
 from hydromt.typing import ErrorHandleMethod, GeoDataframeSource, TotalBounds
 
 from .. import gis_utils, io
+from ..nodata import NoDataStrategy, _exec_nodata_strat
 from .data_adapter import DataAdapter
 
 logger = logging.getLogger(__name__)
@@ -234,6 +235,7 @@ class GeoDataFrameAdapter(DataAdapter):
         buffer=0,
         predicate="intersects",
         logger=logger,
+        handle_nodata=NoDataStrategy.RAISE,
         variables=None,
     ):
         """Return a clipped and unified GeoDataFrame (vector).
@@ -251,7 +253,7 @@ class GeoDataFrameAdapter(DataAdapter):
         gdf = self._set_nodata(gdf)
         # slice
         gdf = GeoDataFrameAdapter._slice_data(
-            gdf, variables, geom, bbox, buffer, predicate, logger=logger
+            gdf, variables, geom, bbox, buffer, predicate, handle_nodata, logger=logger
         )
         # uniformize
         gdf = self._apply_unit_conversions(gdf, logger=logger)
@@ -325,6 +327,7 @@ class GeoDataFrameAdapter(DataAdapter):
         bbox=None,
         buffer=0,
         predicate="intersects",
+        handle_nodata=NoDataStrategy.RAISE,
         logger=logger,
     ):
         """Return a clipped GeoDataFrame (vector).
@@ -340,6 +343,8 @@ class GeoDataFrameAdapter(DataAdapter):
             (in WGS84 coordinates).
         buffer : float, optional
             Buffer around the `bbox` or `geom` area of interest in meters. By default 0.
+        handle_nodata : NoDataStrategy, optional
+            Strategy to handle no data values. By default NoDataStrategy.RAISE.
         predicate : str, optional
             Predicate used to filter the GeoDataFrame, see
             :py:func:`hydromt.gis_utils.filter_gdf` for details.
@@ -365,7 +370,9 @@ class GeoDataFrameAdapter(DataAdapter):
             logger.debug(f"Clip {predicate} [{bbox_str}] (EPSG:{epsg})")
             idxs = gis_utils.filter_gdf(gdf, geom=geom, predicate=predicate)
             if idxs.size == 0:
-                raise IndexError("No data within spatial domain.")
+                _exec_nodata_strat(
+                    "No data within spatial domain.", handle_nodata, logger=logger
+                )
             gdf = gdf.iloc[idxs]
         return gdf