+++ hydromt/data_adapter/rasterdataset.py
@@ -27,6 +27,7 @@ from hydromt.typing import (
 )
 
 from .. import gis_utils, io
+from ..nodata import NoDataStrategy, _exec_nodata_strat
 from ..raster import GEO_MAP_COORD
 from .caching import cache_vrt_tiles
 from .data_adapter import PREPROCESSORS, DataAdapter
@@ -295,7 +296,7 @@ class RasterDatasetAdapter(DataAdapter):
         ds = self._shift_time(ds, logger)
         # slice data
         ds = RasterDatasetAdapter._slice_data(
-            ds, variables, geom, bbox, buffer, align, time_tuple, logger
+            ds, variables, geom, bbox, buffer, align, time_tuple, logger=logger
         )
         # uniformize data
         ds = self._apply_unit_conversions(ds, logger)
@@ -419,6 +420,7 @@ class RasterDatasetAdapter(DataAdapter):
         buffer=0,
         align=None,
         time_tuple=None,
+        handle_nodata=NoDataStrategy.RAISE,
         logger=logger,
     ):
         """Return a RasterDataset sliced in both spatial and temporal dimensions.
@@ -441,6 +443,8 @@ class RasterDatasetAdapter(DataAdapter):
         time_tuple : Tuple of datetime, optional
             A tuple consisting of the lower and upper bounds of time that the
             result should contain
+        handle_nodata: NoDataStrategy, optional
+            How to handle no data values, by default NoDataStrategy.RAISE
 
         Returns
         -------
@@ -464,11 +468,18 @@ class RasterDatasetAdapter(DataAdapter):
             ds = RasterDatasetAdapter._slice_temporal_dimension(
                 ds,
                 time_tuple,
+                handle_nodata,
                 logger=logger,
             )
         if geom is not None or bbox is not None:
             ds = RasterDatasetAdapter._slice_spatial_dimensions(
-                ds, geom, bbox, buffer, align, logger=logger
+                ds,
+                geom,
+                bbox,
+                buffer,
+                align,
+                handle_nodata,
+                logger=logger,
             )
         return ds
 
@@ -487,7 +498,9 @@ class RasterDatasetAdapter(DataAdapter):
         return ds
 
     @staticmethod
-    def _slice_temporal_dimension(ds, time_tuple, logger=logger):
+    def _slice_temporal_dimension(
+        ds, time_tuple, handle_nodata=NoDataStrategy.RAISE, logger=logger
+    ):
         if (
             "time" in ds.dims
             and ds["time"].size > 1
@@ -497,11 +510,21 @@ class RasterDatasetAdapter(DataAdapter):
                 logger.debug(f"Slicing time dim {time_tuple}")
                 ds = ds.sel({"time": slice(*time_tuple)})
                 if ds.time.size == 0:
-                    raise IndexError("Time slice out of range.")
+                    _exec_nodata_strat(
+                        "Time slice out of range.", handle_nodata, logger
+                    )
         return ds
 
     @staticmethod
-    def _slice_spatial_dimensions(ds, geom, bbox, buffer, align, logger=logger):
+    def _slice_spatial_dimensions(
+        ds,
+        geom,
+        bbox,
+        buffer,
+        align,
+        handle_nodata=NoDataStrategy.RAISE,
+        logger=logger,
+    ):
         # make sure bbox is in data crs
         crs = ds.raster.crs
         epsg = crs.to_epsg()  # this could return None
@@ -523,7 +546,11 @@ class RasterDatasetAdapter(DataAdapter):
             logger.debug(f"Clip to [{bbox_str}] (epsg:{epsg}))")
             ds = ds.raster.clip_bbox(bbox, buffer=buffer, align=align)
             if np.any(np.array(ds.raster.shape) < 2):
-                raise IndexError("RasterDataset: No data within spatial domain.")
+                _exec_nodata_strat(
+                    "RasterDataset: No data within spatial domain",
+                    handle_nodata,
+                    logger,
+                )
 
         return ds