+++ hydromt/data_catalog.py
@@ -46,6 +46,8 @@ from .data_adapter import (
     RasterDatasetAdapter,
 )
 from .data_adapter.caching import HYDROMT_DATADIR, _copyfile, _uri_validator
+from .exceptions import NoDataException
+from .nodata import NoDataStrategy
 
 logger = logging.getLogger(__name__)
 
@@ -1179,7 +1181,7 @@ class DataCatalog(object):
                                 time_tuple=time_tuple,
                                 logger=self.logger,
                             )
-                        except IndexError as e:
+                        except NoDataException as e:
                             self.logger.warning(f"{key} file contains no data: {e}")
                             continue
                         # update path & driver and remove kwargs
@@ -1226,6 +1228,7 @@ class DataCatalog(object):
         geom: Optional[gpd.GeoDataFrame] = None,
         zoom_level: Optional[int | tuple] = None,
         buffer: Union[float, int] = 0,
+        handle_nodata=NoDataStrategy.RAISE,
         align: Optional[bool] = None,
         variables: Optional[Union[List, str]] = None,
         time_tuple: Optional[Tuple] = None,
@@ -1313,6 +1316,7 @@ class DataCatalog(object):
                 buffer,
                 align,
                 time_tuple,
+                handle_nodata,
                 logger=self.logger,
             )
             return RasterDatasetAdapter._single_var_as_array(
@@ -1341,6 +1345,7 @@ class DataCatalog(object):
         bbox: Optional[List] = None,
         geom: Optional[gpd.GeoDataFrame] = None,
         buffer: Union[float, int] = 0,
+        handle_nodata=NoDataStrategy.RAISE,
         variables: Optional[Union[List, str]] = None,
         predicate: str = "intersects",
         provider: Optional[str] = None,
@@ -1369,6 +1374,8 @@ class DataCatalog(object):
             A geometry defining the area of interest.
         buffer : float, optional
             Buffer around the `bbox` or `geom` area of interest in meters. By default 0.
+        handle_nodata : NoDataStrategy, optional
+            How to handle no data values, by default NoDataStrategy.RAISE
         predicate : optional
             If predicate is provided, the GeoDataFrame is filtered by testing
             the predicate function against each item. Requires bbox or mask.
@@ -1408,7 +1415,14 @@ class DataCatalog(object):
                 self.add_source(name, source)
         elif isinstance(data_like, gpd.GeoDataFrame):
             return GeoDataFrameAdapter._slice_data(
-                data_like, variables, geom, bbox, buffer, predicate, logger=self.logger
+                data_like,
+                variables,
+                geom,
+                bbox,
+                buffer,
+                predicate,
+                handle_nodata,
+                logger=self.logger,
             )
         else:
             raise ValueError(f'Unknown vector data type "{type(data_like).__name__}"')
@@ -1429,6 +1443,7 @@ class DataCatalog(object):
         bbox: Optional[List] = None,
         geom: Optional[gpd.GeoDataFrame] = None,
         buffer: Union[float, int] = 0,
+        handle_nodata: NoDataStrategy = NoDataStrategy.RAISE,
         predicate: str = "intersects",
         variables: Optional[List] = None,
         time_tuple: Optional[Tuple] = None,
@@ -1509,6 +1524,7 @@ class DataCatalog(object):
                 buffer,
                 predicate,
                 time_tuple,
+                handle_nodata,
                 logger=self.logger,
             )
             return GeoDatasetAdapter._single_var_as_array(
@@ -1533,6 +1549,7 @@ class DataCatalog(object):
         data_like: Union[str, SourceSpecDict, Path, xr.Dataset, xr.DataArray],
         variables: Optional[list] = None,
         time_tuple: Optional[Tuple] = None,
+        handle_nodata: NoDataStrategy = NoDataStrategy.RAISE,
         provider: Optional[str] = None,
         version: Optional[str] = None,
         **kwargs,
@@ -1578,7 +1595,7 @@ class DataCatalog(object):
                 self.add_source(name, source)
         elif isinstance(data_like, pd.DataFrame):
             return DataFrameAdapter._slice_data(
-                data_like, variables, time_tuple, logger=self.logger
+                data_like, variables, time_tuple, handle_nodata, logger=self.logger
             )
         else:
             raise ValueError(f'Unknown tabular data type "{type(data_like).__name__}"')